Stomp.Subscriptions

Description

The Subscriptions module deals with managing subscriptions on the STOMP broker.

Documentation

type ClientId = Integer

A unique client identifier

type Destination = String

A unique destination identifier

data SubscriptionManager

The SubscriptionManager allows the server to add and remove new subscriptions, send messages to destinations, and send ACK/NACK responses.

clientDisconnected :: SubscriptionManager -> ClientId -> IO ()

Report a client disconnect

initManager :: IO SubscriptionManager

Initialize a SubscriptionManager and return it in an IO context.

unsubscribe :: SubscriptionManager -> ClientId -> SubscriptionId -> IO ()

Unsubscribe from a destination.

subscribe :: SubscriptionManager -> Destination -> ClientId -> SubscriptionId -> AckType -> FrameHandler -> IO ()

Subscribe to a destination; if the destination does not already exist it will be created.

sendAckResponse :: SubscriptionManager -> ClientId -> Frame -> IO ()

Send an ack response in an IO context

ackResponseEvt :: SubscriptionManager -> ClientId -> Frame -> Evt ()

Send an ack response in an Evt context

sendMessage :: SubscriptionManager -> Destination -> Frame -> IO ()

Send a Frame to a Destination in an IO context.

sendMessageEvt :: SubscriptionManager -> Destination -> Frame -> Evt ()

Send a Frame to a Destination in an Evt context.